massive refactor to allow for dynamically sized maps in a single world, implemented many optimizations to handle larger maps, added more world management tools

+++ server/src/Mirage.Server.Core/Configuration/ServerConfig.cs
@@ -156,11 +156,12 @@ public sealed record SpawnConfig
     public int Y { get; init; } = (Mirage.Shared.Constants.MaxMapY + 1) / 2;
 
     /// <summary>Where a character comes back to: the point they paid for at an Inn, or this default when
-    /// they have not set one. Death and /home both land here, so neither can drift from the other.</summary>
-    public (short Map, byte X, byte Y) HomeFor(Mirage.Shared.Records.PlayerRecord p) =>
-        p.SpawnMap > 0
-            ? ((short)p.SpawnMap, (byte)p.SpawnX, (byte)p.SpawnY)
-            : ((short)Map, (byte)X, (byte)Y);
+    /// they have not set one. Death and /home both land here, so neither can drift from the other.
+    ///
+    /// <para>Plain integers, so the answer can name any tile on any map. A narrower type here would silently
+    /// wrap a coordinate past its width and put someone somewhere they never chose.</para></summary>
+    public (int Map, int X, int Y) HomeFor(Mirage.Shared.Records.PlayerRecord p) =>
+        p.SpawnMap > 0 ? (p.SpawnMap, p.SpawnX, p.SpawnY) : (Map, X, Y);
 }
 
 /// <summary>
