+++ server/src/Mirage.Server.Core/Net/PacketHandler.Account.cs
@@ -403,8 +403,8 @@ public sealed partial class PacketHandler
         chr.Spd = cls.Spd;
         chr.Int = cls.Int;
         chr.Map = (short)_config.Spawn.Map;
-        chr.X = (byte)_config.Spawn.X;
-        chr.Y = (byte)_config.Spawn.Y;
+        chr.X = _config.Spawn.X;
+        chr.Y = _config.Spawn.Y;
         chr.Hp = StatFormulas.GetPlayerMaxHp(chr, cls);
         chr.Mp = StatFormulas.GetPlayerMaxMp(chr, cls);
         chr.Sp = StatFormulas.GetPlayerMaxSp(chr, cls);
